added error handling middleware, added getByIdArticle

+++ FitnessPortalAPI/Services/ArticleService.cs
@@ -1,4 +1,5 @@
 ﻿using FitnessPortalAPI.Entities;
+using FitnessPortalAPI.Exceptions;
 using FitnessPortalAPI.Models;
 
 namespace FitnessPortalAPI.Services
@@ -7,6 +8,7 @@ namespace FitnessPortalAPI.Services
     {
         int Create(ArticleDto dto);
         List<ArticleDto> GetAll();
+        ArticleDto GetById(int id);
         void Update(int id, UpdateArticleDto dto);
         void RemoveAll();
         void Remove(int articleId);
@@ -57,6 +59,26 @@ namespace FitnessPortalAPI.Services
 
             return articlesDtos;
         }
+        public ArticleDto GetById(int id)
+        {
+            var article = _context
+                .Articles
+                .FirstOrDefault(a => a.Id == id);
+
+            if (article is null)
+                throw new NotFoundException("Article not found");
+
+            var result = new ArticleDto()
+            {
+                Title = article.Title,
+                ShortDescription = article.ShortDescription,
+                Content = article.Content,
+                Category = article.Category,
+                DateOfPublication = article.DateOfPublication,
+            };
+
+            return result;
+        }
 
         public void Update(int id, UpdateArticleDto dto)
         {
@@ -65,50 +87,42 @@ namespace FitnessPortalAPI.Services
                 .FirstOrDefault(a => a.Id == id);
 
             if(article == null)
-            {
-                throw new Exception();
-            }
+                throw new NotFoundException("Article not found");
 
             if(dto.Title != "" && dto.Title != null)
-            {
                 article.Title = dto.Title;
-            }
+            
             if(dto.ShortDescription != "" && dto.ShortDescription != null)
-            {
                 article.ShortDescription = dto.ShortDescription;
-            }
+            
             if(dto.Content != "" && dto.Content != null)
-            {
                 article.Content = dto.Content;
-            }
+            
             if(dto.Category != "" && dto.Category != null) 
-            {
                 article.Category = dto.Category;
-            }
-
-
+            
             _context.SaveChanges();
         }
-
         public void Remove(int articleId)
         {
             var article = _context.Articles.FirstOrDefault(a => a.Id == articleId);
+
             if (article is null)
-            {
-                throw new Exception("Bad Request");
-            }
+                throw new NotFoundException("Article not found");
 
             _context.Remove(article);
             _context.SaveChanges();
         }
-
         public void RemoveAll()
         {
             var articles = _context.Articles.ToList();
+
+            if (articles is null)
+                throw new NotFoundException("Article not found");
+
             _context.RemoveRange(articles);
             _context.SaveChanges();
         }
-
         
     }
 }
